core: move the inspector into its own crate
