+++ src/core/inspect/mod.rs
@@ -1,0 +1,10 @@
+//! # Inspection helpers
+//!
+//! Provides support for inspecting the system, kernel, symbols, etc.
+
+// Re-export inspect.rs
+#[allow(clippy::module_inception)]
+pub(crate) mod inspect;
+pub(crate) use inspect::*;
+
+mod btf;
