+++ src/core/kernel/mod.rs
@@ -1,8 +1,5 @@
 //! # Kernel helpers
 
 // Re-export symbol::Symbol.
-pub(crate) use symbol::Symbol;
-
-mod btf;
-pub(crate) mod inspect;
 pub(crate) mod symbol;
+pub(crate) use symbol::Symbol;
