+++ src/core/mod.rs
@@ -6,6 +6,7 @@
 pub(crate) mod bpf_sys;
 pub(crate) mod events;
 pub(crate) mod filters;
+pub(crate) mod inspect;
 pub(crate) mod kernel;
 pub(crate) mod probe;
 pub(crate) mod signals;
